requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai>=1.50.0
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic>=2.4.1,<2.6
pydantic-settings==2.1.0
asyncpg==0.29.0
pillow==10.2.0

